+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Issue with drifting changes for `fivetran_connector.config.list_strategy` field
 - Issue with re-creation of connectors that are using `destination_schema.prefix` field
 - Supported config fields for `CosmosDB` and `Snowflake DB` connectors
+- Supported `replica_id` in connector config for `MySQL` connectors
 
 ## [0.7.2-pre](https://github.com/fivetran/terraform-provider-fivetran/compare/v0.7.1...v0.7.2-pre)